Footer Message
I did this part for Lucas: I created a footer message with icons indicating content is under a CC-BY-SA license, that we are powered by Drupal, and using git and github.  The copyright notice links to the about page we'll create later.

Of course, at least the name will need to be changed if anyone uses this.