Where Stanley Visits
Before Stanley can be on a map, we need a way to record the locations he visits.

We went to http://example.com/admin/build/modules and enabled the following modules:

* Content
* Number
* Option Widgets
* Text
* Location
* Location Taxonomy
* Node Locations

I'm not sure we'll use all of these, but Lucas wanted to make sure we didn't have to come back to this page right away.

After making the cat get out of the way (cat's can't administer Drupal), we went to http://example.com/admin/content/types and clicked "Add Content Type" to make a content type that will represent Stanley's adventures.

We named the content type "Adventure" and set the type, aka its machine name, to "adventure".

We input "An adventure is a place Stanley goes." as the description, so visitors will know what adding an adventure is all about.

Under submission form settings, we labeled titles as "Adventure name" and body as "What happened".

We left the workflow settings as they were, and moved on to Locative Information.  Under "number of locations", we set that each adventure should have exactly one location (minimum and maximum are both one), and one location at a time should be added.  Under "collection settings", we allowed collection of location name, street location, city, state/province, country, zip code, and via the coordinate chooser.  We didn't collect "additional" after street location -- we don't need to be so particular as to know what unit inside a building Stanley visited, they will probably be all one adventure even if he visits many parts of a location.  We left the rss settings at default, and under "display settings" hid the street location and postal code -- we really only collected those to get the right spot on our map.

We disabled comments by default on Adventures, since I didn't think they would get enough use to bother with.

We clicked "save", and now we have an Adventure content type in which people can log their adventures with Flat Stanley.