Stanley's Map
Now that we have a way to record where Stanley has been, it is time to plot them on a map.

We went to http://example.com/admin/build/modules to enable some more modules:

* GMap
* GMap Location
* Views

I signed in to Google and went to http://code.google.com/apis/maps/signup.html to get a Google map API key, which I entered at http://example.com/admin/settings/gmap

** Note: The API key committed with some versions of this DB is NOT valid.  You have to go get your own, per Google's Terms of Service. **

On that page, we left all the defaults in place until we reached "Map Behavior Flags", where we checked "autozoom" in addition to "nokeyboard", which was already checked by default.  We set the map's default width to 95% and default height to 600px.  Then we skipped to the bottom and clicked "Save Configuration".

At http://example.com/admin/settings/location we made some tweaks to the location module configuration:

* In the "Main Settings" tab, we checked "Use a Google Map to set latitude and longitude" for a pretty map in our location chooser.
* Under the "Map Links" tab, everything was already set to use Google Maps, but we had to go through the "Geocoding Options" tab, set it to address level accuracy, and choose google maps under each country where it was available.

Next, we broke for snacks.  Snacks are an integral part of the Drupal site building process.  After some delicious snacks, we headed to http://example.com/admin/build/views and clicked "Add" to create our adventure map.

We named the view "adventure_map" and gave it the description "Adventure Map".  It is of type "node".

** Note: If you are trying to follow this as a tutorial, consider installing the advanced_help module for lots of useful documentation available right through your Drupal site as you work.  We were lazy and didn't install advanced_help simply because an experienced Drupalista/mom was directing things. **

We added two filters to this view:

* Node: Published (yes)  -- This is so only published adventures show up on the map.
* Node: Type (is one of, Adventure) -- This is so only adventures, not pages or stories, go on our map.

Neither was exposed because we don't need users mucking about with them.

We didn't need any sort criteria: the items will be organized on a map, not listed somewhere.

We added some fields:
* Node:Title -- This is our adventure name.  We left the label blank and checked "Link this field to its node" so people who find it on the map can read the whole adventure.
* Location:Name -- This is the name of the place Stanley visited.
* Location:City -- The city
* Location:Province -- The state or province
* Location:Country -- The country

...this is when we realized it would be a good idea to also note *when* Stanley had an adventure, not just *where* it was.  We'll add that in a bit.

Under "Basic Settings" we set the following:

* Style: GMap
	We clicked the gear next to this setting to access its options.  Here we set the default marker style (Lucas chose blue) and that we wanted to display the node title in a tooltip when someone hovered on its marker.
* Row Style: Fields
	In the options for this setting, we made "city" and "province" inline fields so they would be on the same line, and made the separator ", " so it would look nice.  We also selected to hide empty fields.
* Items to display: Unlimited
	Enter '0' in the limit field to get unlimited.
* Header:
	These are the places Stanley has visited.

We added a page display, since Stanley's map will be its own page.

We gave the page display the path "map" and called its menu item "Map" with the description "Map of Stanley's adventures."  We put it in the Primary links menu.

Then we saved the view.

Lucas was dissapointed to see that the map page didn't have anything on it, so we created an adventure node for Stanley's starting point at our house.