We went to http://example.com/admin/content/node-type/adventure/fields to add a space for pictures of each adventure.  We created a new field with the label "Pictures", the field name "field_pictures" (the "field_" part is pre-filled for us), the type "File" and the form element "Image".  We also used the drag-and-drop to order the parts of the adventure creation form to something more logical.
Choosing "configure" next to our picture field in the list, we came to a page with some more settings.  We set the maximum resolution for images to 1280x800 because Lucas said he might want to use one of them as a desktop background.  We left the file size restrictions blank, so that they would only be limited by the PHP upload size.

Under "Filefield path settings" we chose to put all the pictures of Stanley's adventures in a directory called "adventure-pics".  I set the file alias to "files/pictures/adventures/[filefield-onlyname-original].[filefield-extension-original]" so users wouldn't see the sites/default/... paths.  Then I checked "Enable custom title text" under "title text" settings, set the number of values (under "global settings") to unlimited, and left everything else at the defaults.

When all that was done and saved, we clicked the "Display Fields" tab and excluded the picture field from display in teasers.  That way, if there were many pictures it would not make the teasers huge.

Now users could upload pictures to Stanley's adventures!