Setting up the journal.
First, we enabled the following modules:

* Calendar
* Date API
* Date
* Date Popup
* Date Timezone
* Jquery UI

Then, we went back to http://example.com/admin/content/node-type/adventure/fields and added a date field with a popup calendar to the Adventure content type.

The last step was to go to http://example.com/admin/build/views and activate the "archive" view with a few changes:

* We made the title "Journal"
* We set it to sort by the date field we'd created instead of the date the node was posted.
* We changed the path to "journal"
* We created a menu entry in the Primary Links menu to this view's page called "Journal"

Now our journal is done!